2022.2.10
버그 Fix
1. 카테고리 선택 중에 다른 카테고리 선택시 카테고리 하이라이트가 모두 꺼져버리는 현상 수정
2.  아이템 이동이 안되던 문제 해결
3. 아이템 나누기 기능
3-1 아이템 나눌 때 0개로도 나누어지던 문제 해결
3-2 나눠진 아이템을 같은 아이템이 든 슬롯에 안 놔지던 문제 해결

+++ DeliGrow_GoodLife/Assets/Script/Inventory/InventoryUI.cs
@@ -217,9 +217,8 @@ public class InventoryUI : UI
         m_isMoving = false;
         if (m_allocatedSlot != -1)
         {
-            if(inventory.inventoryDatas[m_allocatedSlot].id == 0)
+            if(inventory.InsertItem(item, m_allocatedSlot))
             {
-                inventory.InsertItem(item, m_allocatedSlot);
                 m_moveSlotImage.gameObject.SetActive(false);
                 m_allocatedSlot = -1;
                 return true;
@@ -261,6 +260,11 @@ public class InventoryUI : UI
     public void ItemDevide(ItemData item)
     {
         Unemphasize();
+        if (item.itemCount==0)
+        {
+            Unemphasize();
+            return;
+        }
         //m_itemDevideWindw
         if(!inventory.DeleteItem(item,m_selectedItemIndex))
         {
